+++ simulation/exp3_linear_payoff.py
