+++ simulation/exp4p_linear_payoff.py
