+++ simulation/linthompsamp_linear_payoff .py
