+++ striatum/bandit/tests/test_exp3.py
