+++ striatum/bandit/tests/test_linthompsamp.py
